Multiple Changes, Read Comment
Fixed text file locations and paths in textfilereader program. Otherwise removed "Characters.java" because it can not be accessed in any packages as it is located in default package. Please use the "GameCharacter.java" class located in "common" package instead. It is a 1 for 1 copy of the previous deleted class.