Added Question Removal/Storage
Added class to handle questions, added new txt files to store the users list of remaining questions and a class that will remove questions asked as they are asked, and based off of the result of the question, so if a question is true then all relevant questions are also removed but if the hair color being black was not true then all other hair color questions remain available. Questions with only 2 answers are removed as well since if one is false then the other option is true.